+++ internal/app/model/menu.go
@@ -2,16 +2,16 @@ package model
 
 type Menu struct {
 	Base
-	Code      string  `json:"code"`
-	Name      string  `json:"name" gorm:"not null;"`
-	Path      string  `json:"path" gorm:"not null;comment:api路径;"`
-	Component string  `json:"component" gorm:"not null;comment:组件路径;" `
-	Icon      string  `json:"icon" gorm:"comment:icon图标;"`
-	Note      string  `json:"note" gorm:"comment:备注;"`
-	Type      int     `json:"type"`
-	Status    *int    `json:"status" gorm:"comment:状态:1正常 2禁用;" `
-	Sort      int     `json:"sort" gorm:"comment:显示顺序;"`
-	ParentId  *uint64 `json:"parent_id" gorm:"column:parent_id"` // 允许为空的父ID
+	Code      string `json:"code"`
+	Name      string `json:"name" gorm:"not null;"`
+	Path      string `json:"path" gorm:"not null;comment:api路径;"`
+	Component string `json:"component" gorm:"not null;comment:组件路径;" `
+	Icon      string `json:"icon" gorm:"comment:icon图标;"`
+	Note      string `json:"note" gorm:"comment:备注;"`
+	Type      int    `json:"type"`
+	Status    *int   `json:"status" gorm:"comment:状态:1正常 2禁用;" `
+	Sort      int    `json:"sort" gorm:"comment:显示顺序;"`
+	ParentId  *int   `json:"parent_id" gorm:"column:parent_id"` // 允许为空的父ID
 
 	Children []Menu `json:"children" gorm:"foreignKey:ParentId"`
 
@@ -19,14 +19,14 @@ type Menu struct {
 }
 
 type MenuReq struct {
-	Code      string  `json:"code"`
-	Name      string  `json:"name" `
-	Path      string  `json:"path"  binding:"required" `
-	Component string  `json:"component"`
-	Icon      string  `json:"icon" `
-	Note      string  `json:"note" `
-	Type      int     `json:"type"  binding:"required"`
-	Status    *int    `json:"status"  `
-	Sort      *int    `json:"sort"`
-	ParentId  *uint64 `json:"parent_id" `
+	Code      string `json:"code"`
+	Name      string `json:"name" `
+	Path      string `json:"path"  binding:"required" `
+	Component string `json:"component"`
+	Icon      string `json:"icon" `
+	Note      string `json:"note" `
+	Type      int    `json:"type"  binding:"required"`
+	Status    *int   `json:"status"  `
+	Sort      *int   `json:"sort"`
+	ParentId  *int   `json:"parent_id" `
 }
